Refactored by organizing more Printer and Driver pagckages
