+++ src/main/java/com/jbt/sysout/PrinterCommon.java
